+++ terminice_core/test/mock_terminal.dart
@@ -1,1 +1,11 @@
-export 'package:terminice_core/src/testing/mock_terminal.dart';
+export 'package:terminice_core/testing.dart'
+    show
+        ErrorTerminal,
+        ErrorTerminalInput,
+        ErrorTerminalOutput,
+        MockTerminal,
+        MockTerminalInput,
+        MockTerminalOutput,
+        SpyTerminal,
+        SpyTerminalInput,
+        SpyTerminalOutput;